requirements.txt
discord>=2.3.2
colorama>=0.4.6
PyYAML>=6.0.1
pytz>=2023.3.post1
aiomysql>=0.2.0